+++ lib/adaptive_library.dart
@@ -9,3 +9,4 @@ export 'package:adaptive_library/src/adaptive_inheritance.dart';
 export 'package:adaptive_library/src/adaptive_progress_indicator.dart';
 export 'package:adaptive_library/src/adaptive_list_tile.dart';
 export 'package:adaptive_library/src/adaptive_switch.dart';
+export 'package:adaptive_library/src/adaptive_textfield.dart';